+++ R/validate_kinship.R
@@ -1,7 +1,9 @@
 #' Validate a kinship matrix
 #'
-#' Tests that the input is a valid kinship matrix (a numeric, square, symmetric R matrix).
+#' Tests that the input is a valid kinship matrix (a numeric, square, and optionally symmetric R matrix).
+#' Intended for matrices to plot and for other uses, including biased estimates, so there is flexibility as to what constitutes a valid kinship matrix.
 #' Throws errors if the input is not as above.
+#' Can instead return `TRUE`/`FALSE` if `logical = TRUE`.
 #' 
 #' True kinship matrices have values strictly between 0 and 1, and diagonal values strictly between 0.5 and 1.
 #' However, estimated matrices may contain values slightly out of range.
@@ -11,8 +13,11 @@
 #' @param sym If `TRUE` (default), the matrix is required to be symmetric.  Otherwise this particular test is skipped.
 #' @param name Default "kinship".
 #' Change to desired variable name for more informative error messages (i.e. "A" when used to validate the `A` matrix inside `popkin_A_min_subpops`).
+#' @param logical If `FALSE` (default), function stops with an error message if the input is not a kinship matrix.
+#' If `TRUE`, function instead returns `TRUE` if the input passed all tests (appears to be a valid kinship matrix) or `FALSE` otherwise.
 #'
-#' @return Nothing
+#' @return If `logical = FALSE` (default), nothing.
+#' If `logical = TRUE`, returns `TRUE` if the input is a valid kinship matrix, `FALSE` otherwise.
 #'
 #' @examples
 #' # this is a valid (positive) example
@@ -42,23 +47,34 @@
 #' # but example passes if we drop symmetry requirement this way
 #' validate_kinship( non_kinship, sym = FALSE )
 #'
+#' # instead of stopping, can get a logical value
+#' # this returns FALSE
+#' validate_kinship( non_kinship, logical = TRUE )
+#'
 #' @export
-validate_kinship <- function(kinship, sym = TRUE, name = 'kinship') {
+validate_kinship <- function(kinship, sym = TRUE, name = 'kinship', logical = FALSE) {
     # die if this is missing
+    # this should die even if `logical = TRUE`!
     if ( missing( kinship ) )
         stop( '`', name, '` matrix is required!' )
+    
     # make sure it is an ordinary matrix or equivalent
     if ( !is.matrix( kinship ) )
-        stop( '`', name, '` must be an R matrix!' )
+        if (logical) return (FALSE) else stop( '`', name, '` must be an R matrix!' )
     # make sure it is numeric
     if ( !is.numeric( kinship ) )
-        stop( '`', name, '` must be numeric!' )
+        if (logical) return (FALSE) else stop( '`', name, '` must be numeric!' )
     # check dimensions
     m <- nrow( kinship )
     n <- ncol( kinship )
     if (n != m)
-        stop( '`', name, '` must be a square matrix!  (nrow ', m, ' != ncol ', n, ')' )
+        if (logical) return (FALSE) else stop( '`', name, '` must be a square matrix!  (nrow ', m, ' != ncol ', n, ')' )
     # test symmetry
     if ( sym && !isSymmetric( kinship ) )
-        stop( '`', name, '` must be a symmetric matrix!' )
+        if (logical) return (FALSE) else stop( '`', name, '` must be a symmetric matrix!' )
+    
+    # if everything passed and we wanted a logical, return TRUE now
+    # don't return anything otherwise
+    if (logical)
+        return (TRUE)
 }